updating & patching improv
